refactor(cloister): preserve file-size ratchets
Issue: PAN-3393
