+++ src/cli/index.ts
@@ -63,8 +63,7 @@ import { killCommand } from './commands/kill.js';
 import { registerResetToPlannedCommand } from './commands/reset-to-planned.js';
 import { pauseCommand } from './commands/pause.js';
 import { unpauseCommand } from './commands/unpause.js';
-import { untroubledCommand } from './commands/untroubled.js';
-import { registerUnstickCommand } from './commands/unstick.js';
+import { untroubledCommand } from './commands/untroubled.js'; import { registerUnstickCommand } from './commands/unstick.js';
 import { forkCommand } from './commands/fork.js';
 import { handoffCommand } from './commands/handoff.js';
 import { unarchiveConversationCommand } from './commands/unarchive-conversation.js';
@@ -417,8 +416,7 @@ program
 program
   .command('untroubled <id>')
   .description('Clear an agent troubled gate without spawning it')
-  .action(untroubledCommand);
-registerUnstickCommand(program);
+  .action(untroubledCommand); registerUnstickCommand(program);
 
 program
   .command('fork [conv]')
